chore: Move ConversationMissingMessagesSystemMessageCellDescription into its own file
