+++ wire-ios/Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ConfigurationMessageCell/Components/ConversationSystemMessageCell.swift
@@ -230,91 +230,6 @@ final class ConversationStartedSystemMessageCellDescription: NSObject, Conversat
 
 }
 
-class ConversationMissingMessagesSystemMessageCellDescription: ConversationMessageCellDescription {
-
-    typealias View = ConversationSystemMessageCell
-    let configuration: View.Configuration
-
-    var message: ZMConversationMessage?
-    weak var delegate: ConversationMessageCellDelegate?
-    weak var actionController: ConversationMessageActionController?
-
-    var showEphemeralTimer: Bool = false
-    var topMargin: Float = 0
-
-    let isFullWidth: Bool = true
-    let supportsActions: Bool = false
-    let containsHighlightableContent: Bool = false
-
-    let accessibilityIdentifier: String? = nil
-    let accessibilityLabel: String?
-
-    init(message: ZMConversationMessage, data: ZMSystemMessageData) {
-        let title = ConversationMissingMessagesSystemMessageCellDescription.makeAttributedString(systemMessageData: data)
-        configuration =  View.Configuration(icon: StyleKitIcon.exclamationMark.makeImage(size: .tiny,
-                                                                                         color: IconColors.foregroundExclamationMarkInSystemMessage),
-                                            attributedText: title,
-                                            showLine: true)
-        accessibilityLabel = title.string
-        actionController = nil
-    }
-
-    private static func makeAttributedString(systemMessageData: ZMSystemMessageData) -> NSAttributedString {
-        let string = localizedString(systemMessageData: systemMessageData)
-        return NSAttributedString(string: string) && UIFont.mediumFont && LabelColors.textDefault
-    }
-
-    private static func localizedString(systemMessageData: ZMSystemMessageData) -> String {
-        typealias Strings = L10n.Localizable.Content.System
-
-        guard !systemMessageData.needsUpdatingUsers else {
-            return Strings.missingMessages
-        }
-
-        let namesOfAddedUsers: [String] = systemMessageData.addedUserTypes.compactMap {
-            guard let user = $0 as? UserType else { return nil }
-            return user.name
-        }.sorted(by: { $0 > $1 })
-
-        let namesOfRemovedUsers: [String] = systemMessageData.removedUserTypes.compactMap {
-            guard let user = $0 as? UserType else { return nil }
-            return user.name
-        }.sorted(by: { $0 > $1 })
-
-        let listOfAddedUsers = ListFormatter.localizedString(byJoining: namesOfAddedUsers)
-        let listOfRemovedUsers = ListFormatter.localizedString(byJoining: namesOfRemovedUsers)
-
-        switch (namesOfAddedUsers.cardinality, namesOfRemovedUsers.cardinality) {
-        case (.zero, .zero):
-            return Strings.missingMessages
-
-        case (.singular, .zero):
-            return Strings.MissingMessages.UsersAdded.singular(listOfAddedUsers)
-
-        case (.plural, .zero):
-            return Strings.MissingMessages.UsersAdded.plural(listOfAddedUsers)
-
-        case (.zero, .singular):
-            return Strings.MissingMessages.UsersRemoved.singular(listOfRemovedUsers)
-
-        case (.zero, .plural):
-            return Strings.MissingMessages.UsersRemoved.plural(listOfRemovedUsers)
-
-        case (.singular, .singular):
-            return Strings.MissingMessages.UsersAddedAndRemoved.singular(listOfAddedUsers, listOfRemovedUsers)
-
-        case (.plural, .singular):
-            return Strings.MissingMessages.UsersAddedAndRemoved.pluralSingular(listOfAddedUsers, listOfRemovedUsers)
-
-        case (.singular, .plural):
-            return Strings.MissingMessages.UsersAddedAndRemoved.singularPlural(listOfAddedUsers, listOfRemovedUsers)
-
-        case (.plural, .plural):
-            return Strings.MissingMessages.UsersAddedAndRemoved.plural(listOfAddedUsers, listOfRemovedUsers)
-        }
-    }
-}
-
 class ConversationIgnoredDeviceSystemMessageCellDescription: ConversationMessageCellDescription {
 
     typealias View = ConversationNewDeviceSystemMessageCell
